refactor(validator): isolate DKG board persistence
